removed comments and cleaned up more code

+++ source/adios2/toolkit/interop/hdf5/HDF5Common.h
@@ -172,30 +172,28 @@ public:
 
     void ReadNativeAttrToIO(core::IO &io, hid_t datasetId, std::string const &pathFromRoot);
 
-    // void SetAdiosStep(int ts);
     void SetAdiosStep(size_t ts);
 
     size_t GetNumAdiosSteps();
     size_t GetAdiosStep() const;
     void WriteAdiosSteps();
 
-    void ReadVariables(unsigned int ts, core::IO &io);
-    void FindVarsFromH5(core::IO &io, hid_t gid, const char *name, const char *heritage,
-                        unsigned int ts);
+    void ReadVariables(size_t ts, core::IO &io);
+    void FindVarsFromH5(core::IO &io, hid_t gid, const char *name, const char *heritage, size_t ts);
     void ReadAllVariables(core::IO &io);
 
     void ReadStringScalarDataset(hid_t dataSetId, std::string &result);
     hid_t GetTypeStringScalar(const std::string &input);
-    void CreateVar(core::IO &io, hid_t h5Type, std::string const &name, unsigned int ts);
+    void CreateVar(core::IO &io, hid_t h5Type, std::string const &name, size_t ts);
 
     template <class T>
-    void AddVar(core::IO &io, std::string const &name, hid_t datasetId, unsigned int ts);
+    void AddVar(core::IO &io, std::string const &name, hid_t datasetId, size_t ts);
 
     // adios only allows a scalar string var
-    void AddSingleString(core::IO &io, std::string const &name, hid_t datasetId, unsigned int ts);
+    void AddSingleString(core::IO &io, std::string const &name, hid_t datasetId, size_t ts);
 
     // decompose array string vars
-    void AddVarString(core::IO &io, std::string const &name, hid_t datasetId, unsigned int ts);
+    void AddVarString(core::IO &io, std::string const &name, hid_t datasetId, size_t ts);
 
     template <class T>
     void AddNonStringAttribute(core::IO &io, std::string const &attrName, hid_t attrId,
@@ -213,7 +211,6 @@ public:
     hid_t m_DefH5TypeComplexFloat;
     hid_t m_DefH5TypeBlockStat;
 
-    // unsigned int m_CurrentAdiosStep = 0;
     size_t m_CurrentAdiosStep = 0;
 
     void CheckWriteGroup();
@@ -249,7 +246,7 @@ private:
                           std::vector<hsize_t> &, std::vector<hsize_t> &);
 
     bool m_WriteMode = false;
-    // unsigned int m_NumAdiosSteps = 0;
+
     size_t m_NumAdiosSteps = 0;
 
     MPI_API const *m_MPI = nullptr;
